split cpu_fusion into cpu_fusion and cpu_test; clean up headers
